Move public-facing files to the api folder
In this way users of the uVisor library can find the relevant header
files more quickly, in a unique place.

Header file inclusion has been fixed as well.
