+++ api/inc/register_gateway.h
