+++ api/inc/secure_access.h
