+++ api/inc/secure_gateway.h
