+++ api/src/debug.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015, ARM Limited, All Rights Reserved
+ * Copyright (c) 2016, ARM Limited, All Rights Reserved
  * SPDX-License-Identifier: Apache-2.0
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may
@@ -14,9 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "uvisor-lib/uvisor-lib.h"
+#include "api/inc/uvisor-lib.h"
 
-void uvisor_error(THaltUserError reason)
+UVISOR_EXTERN void uvisor_debug_init(const TUvisorDebugDriver * const driver)
 {
-    UVISOR_SVC(UVISOR_SVC_ID_HALT_USER_ERR, "", reason);
+    UVISOR_SVC(UVISOR_SVC_ID_DEBUG_REGISTER_BOX, "", (uint32_t) driver);
 }
